Widen canonical ChannelHash from u32 to u64
PermissionToken's channel_hash field, TokenCache's fast-path key, the
ChannelConfig registry by-hash index, the AuthGuard bloom key, and
the FFI net_channel_hash output all used a u32 truncation of
xxh3_64. ~2^32 work to find a targeted collision is feasible offline
for a motivated attacker — a colliding channel name lets a token
issued for the grinded name pass TokenCache::check against an
unrelated victim channel that hashes to the same u32 bucket. The
AuthGuard::is_authorized_full path was unaffected (keyed on the full
name string), but every hot-path consumer goes through the hash-only
cache.

Widen ChannelHash to u64 (full xxh3_64 keyspace). The targeted-
collision cost rises to ~2^64. Token wire format channel_hash field
grows from 4 to 8 bytes; WIRE_SIZE 161 → 165 and SIGNED_PAYLOAD_SIZE
97 → 101. The 64-bit signature offset shifts. Bloom-key buffer grows
from 12 to 16 bytes. FFI net_channel_hash now writes through
*mut u64 — bindings that hard-coded u32 must be updated alongside
this commit (the bindings consume the JSON-form parse_token output,
which uses u64 already for not_before / not_after / nonce, so the
NAPI / PyO3 / Go pass-through paths handle the widened type without
code change).

This is a wire-format break: tokens serialized under the old 161-
byte format do not parse under the new 165-byte WIRE_SIZE check.
Acceptable at pre-1.0; tokens are short-lived (issuer-set TTL,
typically minutes to days).

Regression test pins the u32-aliased-u64-hash distinction: a token
bound to hash h_a (= 0x0000_0001_DEAD_BEEF) is rejected against an
h_b (= 0xDEAD_BEEF_DEAD_BEEF) that aliased h_a in the prior u32
keyspace.

+++ net/crates/net/src/ffi/mesh.rs
@@ -2561,12 +2561,12 @@ pub extern "C" fn net_delegate_token(
     }
 }
 
-/// Hash a channel name to its canonical 32-bit [`ChannelHash`]
+/// Hash a channel name to its canonical 64-bit [`ChannelHash`]
 /// (substrate-wide ACL / config / storage key). The 16-bit wire
 /// hash used by `NetHeader::channel_hash` is the low 16 bits of
 /// the returned value. Returns `NET_ERR_IDENTITY` for invalid names.
 #[unsafe(no_mangle)]
-pub extern "C" fn net_channel_hash(channel: *const c_char, out_hash: *mut u32) -> c_int {
+pub extern "C" fn net_channel_hash(channel: *const c_char, out_hash: *mut u64) -> c_int {
     if channel.is_null() || out_hash.is_null() {
         return NetError::NullPointer.into();
     }
